Subscribe to Comments 2.0.2

+++ wp-content/plugins/subscribe-to-comments-reloaded/uninstall.php
@@ -1,11 +1,16 @@
 <?php
 
+// Avoid misusage
+if (!defined('WP_UNINSTALL_PLUGIN')) exit;
+
 global $wpdb;
 
 // Goodbye data...
-$wpdb->query("DROP TABLE IF EXISTS `{$wpdb->prefix}subscribe_reloaded`");
+$wpdb->query("DROP TABLE IF EXISTS {$wpdb->prefix}subscribe_reloaded"); // Compatibility with versions prior to 1.7
+$wpdb->query("DELETE FROM $wpdb->postmeta WHERE meta_key LIKE '\_stcr@\_%'");
 
 // Goodbye options...
+delete_option('subscribe_reloaded_manager_page_enabled');
 delete_option('subscribe_reloaded_manager_page');
 delete_option('subscribe_reloaded_manager_page_title');
 delete_option('subscribe_reloaded_request_mgmt_link');
@@ -16,6 +21,8 @@ delete_option('subscribe_reloaded_subscription_confirmed_dci');
 delete_option('subscribe_reloaded_author_text');
 delete_option('subscribe_reloaded_user_text');
 
+delete_option('subscribe_reloaded_show_subscription_box');
+delete_option('subscribe_reloaded_enable_advanced_subscriptions');
 delete_option('subscribe_reloaded_purge_days');
 delete_option('subscribe_reloaded_from_name');
 delete_option('subscribe_reloaded_from_email');
@@ -26,11 +33,12 @@ delete_option('subscribe_reloaded_enable_html_emails');
 delete_option('subscribe_reloaded_process_trackbacks');
 delete_option('subscribe_reloaded_enable_admin_messages');
 delete_option('subscribe_reloaded_admin_subscribe');
+delete_option('subscribe_reloaded_custom_header_meta');
 
 delete_option('subscribe_reloaded_notification_subject');
 delete_option('subscribe_reloaded_notification_content');
 delete_option('subscribe_reloaded_checkbox_label');
-delete_option('subscribe_reloaded_checkbox_class');
+delete_option('subscribe_reloaded_checkbox_class'); // Not used in StCR >= 2.0, but left for backward compatibility
 delete_option('subscribe_reloaded_checkbox_inline_style');
 delete_option('subscribe_reloaded_checkbox_html');
 delete_option('subscribe_reloaded_subscribed_label');
